update on product and user controller

+++ src/controllers/productController.js
@@ -3,26 +3,20 @@ const AppError = require('../utils/appError');
 const Product = require('../modals/productModels');
 const catchAsync = require('../utils/catchAsync');
 
-
-const reqBody = req => ({ 
-    price: req.body.price,
-    size: req.body.size,
-    ratingsAverage: req.body.ratingsAverage,
-    ratingsQuantity: req.body.ratingsQuantity,
-    productGallery: req.files.map(path => path.path),
-    // productGallery: req.file.path,
-    name: req.body.name,
-    description: req.body.description,
-    brand: req.body.name,
-    category: req.body.category,
-    collectionsData: req.body.collectionsData
-});
-
-
 //CREATE PRODUCT CONTROLLER
 exports.createProduct = catchAsync(async (req, res, next) => {
-    const newProduct = await Product.create(reqBody(req));
-    await newProduct.save({ validateBeforeSave: false });
+    const { price, size, name, description, brand, category, collectionsData } = req.body;
+    const productGallery = req.files ? req.files.map(path => path.path) : [];
+    const newProduct = await Product.create({
+        price,
+        size,
+        productGallery,
+        name,
+        description,
+        brand,
+        category,
+        collectionsData
+    });
 
     res.status(201).json({
         status: 'success',
@@ -31,6 +25,7 @@ exports.createProduct = catchAsync(async (req, res, next) => {
         }
     });
 });
+
 
 
 // GET SINGLE PRODUCT CONTROLLER
@@ -67,26 +62,43 @@ exports.getAllProduct = catchAsync(async (req, res, next) => {
 
 // UPDATE SINGLE PRODUCT CONTROLLER
 exports.updateProduct = catchAsync(async (req, res, next) => {
+    const { price, size, name, description, brand, category, collectionsData } = req.body;
+    const productGallery = req.files ? req.files.map(path => path.path) : [];
     //Updating Image In Uploads File
     const image = await Product.findById(req.params.id);
     const image1 = image.productGallery[0];
     const image2 = image.productGallery[1];
+    const image3 = image.productGallery[2];
+    const image4 = image.productGallery[3];
 
+    // console.log(image);
     if (!image) {
         return next(new AppError('No product found with that ID', 404));
     }
+
     let resultHandler = function (err) {
         if (err) {
             console.log(err);
         }
     }
 
-    fs.unlink(`${image1}`, resultHandler);
-    fs.unlink(`${image2}`, resultHandler);
+    if (image1) { fs.unlink(`${image1}`, resultHandler); }
+    if (image2) { fs.unlink(`${image2}`, resultHandler); }
+    if (image3) { fs.unlink(`${image3}`, resultHandler); }
+    if (image4) { fs.unlink(`${image4}`, resultHandler); }
 
     //Updating product from database
     const product = await Product.findByIdAndUpdate(req.params.id,
-        reqBody(req), {
+        ({
+            price,
+            size,
+            productGallery,
+            name,
+            description,
+            brand,
+            category,
+            collectionsData
+        }), {
             new: true,
             runValidator: true
         });
@@ -94,13 +106,13 @@ exports.updateProduct = catchAsync(async (req, res, next) => {
     if (!product) {
         return next(new AppError('No product found with that ID', 404));
     };
+
     res.status(200).json({
         status: 'success',
         data: {
             product
         }
     });
-
 });
